Functions renamed in phsmethods 0.2.1 (Dec 2021)

phsmethods 0.2.1 renamed a number of functions to improve code clarity. The old functions no longer work and will error. At the next update they will be removed completely.

postcode() -> format_postcode()

age_group() -> create_age_groups()

fin_year() -> extract_fin_year()

Usage

postcode(x, format = c("pc7", "pc8"))

age_group(x, from = 0, to = 90, by = 5, as_factor = FALSE)

fin_year(date)